Split a test class.
